requirements.txt
Flask==2.3.2
python-dotenv==0.21.0
requests==2.28.1
gunicorn==20.1.0
flask-cors==3.0.10